Refactor: Separate jsonToData function
