+++ src/routes/HealthCare/Chart/CompareScoreChart/jsonToChartData.ts
@@ -1,0 +1,16 @@
+import json from 'assets/json/response.json'
+
+const WHSCORE = json.wxcResultMap.wHscore
+const WHSCORYDY = json.wxcResultMap.wHscoreDy.replace(/\[|\]/g, '').split(', ').at(-1)
+
+interface Score {
+  indicator: string
+  score: number
+}
+
+const chartData: Score[] = [
+  { indicator: '나', score: Number(WHSCORE) },
+  { indicator: '10년 후', score: Number(WHSCORYDY) },
+]
+
+export default chartData
